Pre-generate DT decision plot.
Running the script is very memory intensive and crashes on the GKE
binder machines.

Furthermore, including the resulting figure in the markdown is nicer
to reduce the cognitive overload of parsing the "%run script.py"
syntax that early in the learning process of the mooc for people
not too familiar with Python and Jupyter.